web/fix: redirect hocs
