+++ packages/web/app/_hoc/with-public-only-page.tsx
@@ -1,0 +1,19 @@
+import { FC } from 'react';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/domain/auth';
+import { redirect } from '@/navigation';
+import { paths } from '@/navigation/paths';
+
+type Props = {
+  params: { locale: string };
+};
+
+export const withPublicOnlyPage = (Component: FC<Props>) => {
+  return async (props: Props) => {
+    const session = await getServerSession(authOptions);
+    if (session) {
+      redirect(paths.home());
+    }
+    return <Component {...props} />;
+  };
+};
